+++ app/src/androidTest/java/com/college/converter/suntest/SunActivityTest006.java
@@ -1,4 +1,4 @@
-package com.college.converter;
+package com.college.converter.suntest;
 
 
 import static androidx.test.espresso.Espresso.onView;
@@ -8,7 +8,6 @@ import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static androidx.test.espresso.matcher.ViewMatchers.withContentDescription;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static androidx.test.espresso.matcher.ViewMatchers.withParent;
-import static androidx.test.espresso.matcher.ViewMatchers.withText;
 import static org.hamcrest.Matchers.allOf;
 
 import android.view.View;
@@ -16,35 +15,34 @@ import android.view.ViewGroup;
 import android.view.ViewParent;
 
 import androidx.test.espresso.ViewInteraction;
+import androidx.test.espresso.matcher.ViewMatchers;
 import androidx.test.ext.junit.rules.ActivityScenarioRule;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.filters.LargeTest;
 
+import com.college.converter.MainActivity;
+import com.college.converter.R;
+
 import org.hamcrest.Description;
 import org.hamcrest.Matcher;
+import org.hamcrest.Matchers;
 import org.hamcrest.TypeSafeMatcher;
-import org.hamcrest.core.IsInstanceOf;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
 @LargeTest
 @RunWith(AndroidJUnit4.class)
-public class SunActivityTest001 {
+public class SunActivityTest006 {
 
     @Rule
     public ActivityScenarioRule<MainActivity> mActivityScenarioRule =
             new ActivityScenarioRule<>(MainActivity.class);
 
     @Test
-    public void sunActivityTest001() {
-        try {
-            Thread.sleep(5929);
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
+    public void sunActivityTest006() {
         ViewInteraction actionMenuItemView = onView(
-                allOf(withId(R.id.first_id), withContentDescription("Sunrise/Sunset"),
+                Matchers.allOf(ViewMatchers.withId(R.id.first_id), withContentDescription("Sunrise/Sunset"),
                         childAtPosition(
                                 childAtPosition(
                                         withId(R.id.mainToolbar),
@@ -52,17 +50,12 @@ public class SunActivityTest001 {
                                 1),
                         isDisplayed()));
         actionMenuItemView.perform(click());
-        try {
-            Thread.sleep(5929);
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
-        ViewInteraction editText = onView(
-                allOf(withId(R.id.editText_latitude), withText("0"),
-                        withParent(allOf(withId(R.id.linear_layout),
-                                withParent(IsInstanceOf.<View>instanceOf(android.view.ViewGroup.class)))),
+
+        ViewInteraction button = onView(
+                allOf(withId(R.id.help), withContentDescription("Help"),
+                        withParent(withParent(withId(R.id.toolbar))),
                         isDisplayed()));
-        editText.check(matches(withText("0")));
+        button.check(matches(isDisplayed()));
     }
 
     private static Matcher<View> childAtPosition(
